+++ spec/togls/rules/boolean_spec.rb
@@ -1,6 +1,18 @@
 require 'spec_helper'
 
 describe Togls::Rules::Boolean do
+  describe '.title' do
+    it 'does not raise an error' do
+      Togls::Rules::Boolean.title
+    end
+  end
+
+  describe '.description' do
+    it 'does not raise an error' do
+      Togls::Rules::Boolean.description
+    end
+  end
+
   describe "#run" do
     it "returns the provided boolean value" do
       bool_rule = Togls::Rules::Boolean.new(true)