+++ spec/togls/rules/group_spec.rb
@@ -1,6 +1,18 @@
 require 'spec_helper'
 
 describe Togls::Rules::Group do
+  describe '.title' do
+    it 'does not raise an error' do
+      Togls::Rules::Boolean.title
+    end
+  end
+
+  describe '.description' do
+    it 'does not raise an error' do
+      Togls::Rules::Boolean.description
+    end
+  end
+
   describe "#run" do
     context "when the target is included in the list" do
       it "returns true" do